refactor: remove tests from main.py
